+++ changes.md
@@ -1,5 +1,8 @@
 # Recent changes
 
+## 21 June 2023
+- Automatically create new Additional Style Files if needed
+
 ## 20 June 2023
 - Regular backups created in `extensions/Styles-Editor/backups` (saved every ten minutes if changes have been made, last 12 retained)
 - Updated most of the documentation